+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
-- npm publishing logic
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
+## [0.3.0-1] - 2020-08-27
+### Changed
+- npm publishing logic
 ## [0.3.0-0] - 2020-08-27
 ### Breaking changes
 - made `BrowserLogger` a `LogFunctionFactory` (`() => LogFunction`) for consistency.
@@ -61,9 +63,10 @@ ChalkLogger(chalk) // generates a new LogFunction using the chalk object
 ## v0.1.1 - 2020-08-26
 Initial prerelease.
 
-[Unreleased]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-0...HEAD
+[Unreleased]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-1...HEAD
 [0.1.2]: https://github.com/fliegwerk/logsemts/compare/v0.1.1...v0.1.2
 [0.1.3]: https://github.com/fliegwerk/logsemts/compare/v0.1.2...v0.1.3
 [0.1.4]: https://github.com/fliegwerk/logsemts/compare/v0.1.3...v0.1.4
 [0.2.0]: https://github.com/fliegwerk/logsemts/compare/v0.1.4...v0.2.0
 [0.3.0-0]: https://github.com/fliegwerk/logsemts/compare/v0.2.0...v0.3.0-0
+[0.3.0-1]: https://github.com/fliegwerk/logsemts/compare/v0.3.0-0...v0.3.0-1